+++ ci4/app/Controllers/Deployments.php
@@ -119,7 +119,7 @@ class Deployments extends ResourceController {
                 $item->updateIngress(
                     $this->request->getGet('domainId'),
                     $this->request->getGet('subdomain'),
-                    $this->request->getGet('aliases')
+                    $this->request->getGet('aliases') ?? ''
                 );
             } catch (ValidationException $e) {
                 $this->fail($e->getMessage());